MAINT: Refactor settings and file chooser in gui
